Revert "Styling services and services info"
This reverts commit a3a33c41a4ca95a8893daab4f037ff845f3994e8.

+++ XadamaWebapp/lib/CAD/cadride.cs
@@ -75,8 +75,8 @@ namespace lib.CAD
                 ride.name = t.Rows[0][1].ToString();
                 ride.zone = t.Rows[0][2].ToString();
                 ride.description = t.Rows[0][3].ToString();
-                ride.minHeight =  float.Parse(t.Rows[0][4].ToString());
-                ride.speed = float.Parse(t.Rows[0][5].ToString());
+                ride.minHeight = Int32.Parse(t.Rows[0][4].ToString());
+                ride.speed = Int32.Parse(t.Rows[0][5].ToString());
                 ride.image = t.Rows[0][6].ToString();
 
             }
